Moved gui from main.py to gui module
