+++ visma/gui/cli.py
@@ -1,0 +1,2 @@
+def commandExec(command):
+    pass
